+++ packages/core/src/graph/controller/layout.ts
@@ -66,12 +66,12 @@ export default abstract class LayoutController {
 
   // 绘制
   public refreshLayout() {
-    const { graph } = this;
+    const { graph, layoutType, data } = this;
     if (!graph) return;
     if (graph.get('animate')) {
-      graph.positionsAnimate();
+      graph.positionsAnimate(layoutType === 'comboCombined');
     } else {
-      graph.refreshPositions();
+      graph.refreshPositions(layoutType === 'comboCombined');
     }
   }
 
@@ -154,7 +154,7 @@ export default abstract class LayoutController {
       if (comboItem.destroyed) continue;
       const model = comboItem.getModel();
       if (!comboItem.isVisible()) {
-        hiddenEdges.push(model);
+        hiddenCombos.push(model);
         continue;
       }
       combos.push(model);
@@ -214,13 +214,15 @@ export default abstract class LayoutController {
       start = start.then(() => this.reLayoutMethod(layoutMethod, currentCfg));
     });
 
-    start
-      .then(() => {
-        if (layoutCfg.onAllLayoutEnd) layoutCfg.onAllLayoutEnd();
-      })
-      .catch((error) => {
-        console.warn('relayout failed', error);
-      });
+    if (layoutMethods?.length) {
+      start
+        .then(() => {
+          if (layoutCfg.onAllLayoutEnd) layoutCfg.onAllLayoutEnd();
+        })
+        .catch((error) => {
+          console.warn('relayout failed', error);
+        });
+    }
   }
 
   // 筛选参与布局的nodes和edges
@@ -279,7 +281,7 @@ export default abstract class LayoutController {
 
   // 控制布局动画
   // eslint-disable-next-line class-methods-use-this
-  public layoutAnimate() { }
+  public layoutAnimate() {}
 
   // 将当前节点的平均中心移动到原点
   public moveToZero() {