+++ packages/core/src/graph/graph.ts
@@ -613,7 +613,6 @@ export default abstract class AbstractGraph extends EventEmitter implements IAbs
       });
     } else {
       matrix = transform(matrix, [['t', dx, dy]]);
-
       group.setMatrix(matrix);
 
       this.emit('viewportchange', { action: 'translate', matrix });
@@ -1005,7 +1004,7 @@ export default abstract class AbstractGraph extends EventEmitter implements IAbs
     if (isString(item)) nodeItem = this.findById(item as string);
 
     if (!nodeItem && isString(item)) {
-      console.warn('The item to be removed does not exist!');
+      console.warn(`The item ${item} to be removed does not exist!`);
     } else if (nodeItem) {
       let type = '';
       if ((nodeItem as Item).getType) type = (nodeItem as Item).getType();
@@ -1095,8 +1094,7 @@ export default abstract class AbstractGraph extends EventEmitter implements IAbs
     }
 
     let item;
-    let comboTrees = this.get('comboTrees');
-    if (!comboTrees) comboTrees = [];
+    const comboTrees = this.get('comboTrees') || [];
     if (type === 'combo') {
       const itemMap = this.get('itemMap');
       let foundParent = false;
@@ -1148,7 +1146,7 @@ export default abstract class AbstractGraph extends EventEmitter implements IAbs
       const itemMap = this.get('itemMap');
       let foundParent = false,
         foundNode = false;
-      (comboTrees || []).forEach((ctree: ComboTree) => {
+      comboTrees.forEach((ctree: ComboTree) => {
         if (foundNode || foundParent) return; // terminate the forEach
         traverseTreeUp<ComboTree>(ctree, child => {
           if (child.id === model.id) {
@@ -1381,8 +1379,8 @@ export default abstract class AbstractGraph extends EventEmitter implements IAbs
     });
 
     // process the data to tree structure
-    if (combos && combos.length !== 0) {
-      const comboTrees = plainCombosToTrees(combos, nodes);
+    if (combos?.length !== 0) {
+      const comboTrees = plainCombosToTrees((combos as ComboConfig[]), (nodes as NodeConfig[]));
       this.set('comboTrees', comboTrees);
       // add combos
       self.addCombos(combos);
@@ -1725,9 +1723,9 @@ export default abstract class AbstractGraph extends EventEmitter implements IAbs
         if (subtree.id === comboId) {
           treeToBeUncombo = subtree;
           // delete the related edges
-          const edges = comboItem.getEdges();
-          edges.forEach(edge => {
-            this.removeItem(edge, false);
+          const edgeIds = comboItem.getEdges().map(edge => edge.getID());
+          edgeIds.forEach(edgeId => {
+            this.removeItem(edgeId, false);
           });
           const index = comboItems.indexOf(comboItem);
           comboItems.splice(index, 1);
@@ -1783,9 +1781,10 @@ export default abstract class AbstractGraph extends EventEmitter implements IAbs
   }
 
   /**
-   * 根据节点的 bbox 更新所有 combos 的绘制，包括 combos 的位置和范围
+   * 根据 combo 位置更新内部节点位置 followCombo = true
+   * 或根据内部元素的 bbox 更新所有 combos 的绘制，包括 combos 的位置和范围，followCombo = false
    */
-  public updateCombos() {
+  public updateCombos(followCombo: boolean = false) {
     const self = this;
     const comboTrees = this.get('comboTrees');
     const itemController: ItemController = self.get('itemController');
@@ -1803,7 +1802,7 @@ export default abstract class AbstractGraph extends EventEmitter implements IAbs
           each(states, state => this.setItemState(childItem, state, false));
 
           // 更新具体的 Combo
-          itemController.updateCombo(childItem, child.children);
+          itemController.updateCombo(childItem, child.children, followCombo);
 
           // 更新 Combo 后，还原已有的状态
           each(states, state => this.setItemState(childItem, state, true));
@@ -2113,7 +2112,7 @@ export default abstract class AbstractGraph extends EventEmitter implements IAbs
   /**
    * 根据 graph 上的 animateCfg 进行视图中节点位置动画接口
    */
-  public positionsAnimate(): void {
+  public positionsAnimate(referComboModel?: boolean): void {
     const self = this;
     self.emit('beforeanimate');
 
@@ -2121,7 +2120,7 @@ export default abstract class AbstractGraph extends EventEmitter implements IAbs
 
     const { onFrame } = animateCfg;
 
-    const nodes = self.getNodes();
+    const nodes = referComboModel ? self.getNodes().concat(self.getCombos()) : self.getNodes();
 
     const toNodes = nodes.map(node => {
       const model = node.getModel();
@@ -2151,26 +2150,31 @@ export default abstract class AbstractGraph extends EventEmitter implements IAbs
 
           const model: NodeConfig = node.get('model');
 
-          if (!originAttrs) {
-            let containerMatrix = node.getContainer().getMatrix();
-            if (!containerMatrix) containerMatrix = [1, 0, 0, 0, 1, 0, 0, 0, 1];
-            originAttrs = {
+          let containerMatrix = node.getContainer().getMatrix();
+
+          if (originAttrs === undefined) {
+            // 变换前存在位置，设置到 originAttrs 上。否则标记 0 表示变换前不存在位置，不需要计算动画
+            node.set('originAttrs', containerMatrix ? {
               x: containerMatrix[6],
               y: containerMatrix[7],
-            };
-            node.set('originAttrs', originAttrs);
+            } : 0);
           }
 
           if (onFrame) {
-            const attrs = onFrame(node, ratio, data, originAttrs);
+            const attrs = onFrame(node, ratio, data, originAttrs || { x: 0, y: 0 });
             node.set('model', Object.assign(model, attrs));
-          } else {
+          } else if (originAttrs) {
+            // 变换前存在位置，进行动画
             model.x = originAttrs.x + (data.x - originAttrs.x) * ratio;
             model.y = originAttrs.y + (data.y - originAttrs.y) * ratio;
+          } else {
+            // 若在变换前不存在位置信息，则直接放到最终位置上
+            model.x = data.x;
+            model.y = data.y;
           }
         });
 
-        self.refreshPositions();
+        self.refreshPositions(referComboModel);
       },
       {
         duration: animateCfg.duration,
@@ -2193,7 +2197,7 @@ export default abstract class AbstractGraph extends EventEmitter implements IAbs
   /**
    * 当节点位置在外部发生改变时，刷新所有节点位置，重计算边
    */
-  public refreshPositions() {
+  public refreshPositions(referComboModel?: boolean) {
     const self = this;
     self.emit('beforegraphrefreshposition');
 
@@ -2206,19 +2210,26 @@ export default abstract class AbstractGraph extends EventEmitter implements IAbs
 
     const updatedNodes: { [key: string]: boolean } = {};
 
-    each(nodes, (node: INode) => {
-      model = node.getModel() as NodeConfig;
-      const originAttrs = node.get('originAttrs');
-      if (originAttrs && model.x === originAttrs.x && model.y === originAttrs.y) {
-        return;
-      }
-      const changed = node.updatePosition({ x: model.x!, y: model.y! });
-      updatedNodes[model.id] = changed;
-      if (model.comboId) updatedNodes[model.comboId] = updatedNodes[model.comboId] || changed;
-    });
+    const updateItems = (items) => {
+      each(items, (item: INode) => {
+        model = item.getModel() as NodeConfig;
+        const originAttrs = item.get('originAttrs');
+        if (originAttrs && model.x === originAttrs.x && model.y === originAttrs.y) {
+          return;
+        }
+        const changed = item.updatePosition({ x: model.x!, y: model.y! });
+        updatedNodes[model.id] = changed;
+        if (model.comboId) updatedNodes[model.comboId] = updatedNodes[model.comboId] || changed;
+      });
+    }
+
+    updateItems(nodes);
 
     if (combos && combos.length !== 0) {
       self.updateCombos();
+      if (referComboModel) {
+        updateItems(combos);
+      }
     }
 
     each(edges, (edge: IEdge) => {
@@ -2398,6 +2409,7 @@ export default abstract class AbstractGraph extends EventEmitter implements IAbs
    * @param {string | ICombo} combo combo ID 或 combo item
    */
   public collapseCombo(combo: string | ICombo): void {
+    if (this.destroyed) return;
     if (isString(combo)) {
       combo = this.findById(combo) as ICombo;
     }