+++ packages/pc/src/layout/index.ts
@@ -6,6 +6,7 @@ import {
   ForceLayout,
   CircularLayout,
   DagreLayout,
+  DagreCompoundLayout,
   RadialLayout,
   ConcentricLayout,
   MDSLayout,
@@ -14,8 +15,9 @@ import {
   GForceLayout,
   GForceGPULayout,
   ComboForceLayout,
-  ForceAtlas2Layout
-} from '@antv/layout';
+  ComboCombinedLayout,
+  ForceAtlas2Layout,
+} from '@antv/layout/lib';
 
 import TreeLayout from './tree-layout';
 
@@ -24,6 +26,7 @@ oRegisterLayout('random', RandomLayout);
 oRegisterLayout('force', ForceLayout);
 oRegisterLayout('circular', CircularLayout);
 oRegisterLayout('dagre', DagreLayout);
+oRegisterLayout('dagreCompound', DagreCompoundLayout);
 oRegisterLayout('radial', RadialLayout);
 oRegisterLayout('concentric', ConcentricLayout);
 oRegisterLayout('mds', MDSLayout);
@@ -32,6 +35,7 @@ oRegisterLayout('fruchterman-gpu', FruchtermanGPULayout);
 oRegisterLayout('gForce', GForceLayout);
 oRegisterLayout('gForce-gpu', GForceGPULayout);
 oRegisterLayout('comboForce', ComboForceLayout);
+oRegisterLayout('comboCombined', ComboCombinedLayout);
 oRegisterLayout('forceAtlas2', ForceAtlas2Layout);
 
 const registerLayout = (name: string, layoutOverride: any) => {