+++ packages/pc/tests/unit/element/nodes/icon-iconfont-spec.ts
@@ -14,7 +14,7 @@ describe('icon with iconfont', () => {
     },
   };
   const graph = new Graph(cfg);
-  it('default circle config', () => {
+  it.only('default circle config', () => {
     const data = {
       nodes: [
         {
@@ -37,7 +37,6 @@ describe('icon with iconfont', () => {
       })
     })
 
-    graph.emit('canvas:click', {});
     expect(graph.getNodes()[0].get('group').find(e => e.get('name') === 'circle-icon').attr('text')).toBe('xxx');
   });
   it('update iconfont node', () => {